Smoke test, memory, apiTest (#27)
* Smoke test, memory, apiTest

Created Dummy files, smoke test, and memory testing

* Adjusted types and changed mokito syntax

* Adjusted type parameters

* more parameter adjusting

* Assignment3 submission

Will commit folder re-arrangement later

* CREATED SMOKE TESTS!

* Checktyle fix

* removed unnessasry files for this build

+++ src/apis/DataBaseHandler.java
@@ -2,13 +2,13 @@ package apis;
 public interface DataBaseHandler{
 
 
-    public String storeData(long l);
-    //store long l into something? maybe a file
+    public String storeData(int l);
+    //store int l into something? maybe a file
     //send message to successfully store data
 
-    public long sendData(String s);
+    public int sendData(String s);
     //accepts the location of a file to process
 
-
+    
 
 }
